Moved tests into packages
